refactor: 旧LandingPageの未使用実装を削除
