--- src/components/features/LandingPage/faqs.ts
@@ -1,35 +1,0 @@
-export type Faq = {
-  q: string;
-  a: string;
-};
-
-export const faqs: Faq[] = [
-  {
-    q: "LINEでシフト提出はできますか？",
-    a: "できます。スタッフにシフトリから提出用リンクを送り、LINEまたはメールから希望を提出してもらえます。",
-  },
-  {
-    q: "無料で使えるシフト管理ツールですか？",
-    a: "無料のShift管理ツールです。",
-  },
-  {
-    q: "スタッフはアプリ登録が必要ですか？",
-    a: "不要です。届いたリンクを開き、スマホから希望シフトを提出できます。",
-  },
-  {
-    q: "LINEを使わないスタッフがいても利用できますか？",
-    a: "利用できます。LINE未連携のスタッフにはメールで通知します。",
-  },
-  {
-    q: "スマホでもシフト作成できますか？",
-    a: "スマホでも確認・作成できます。細かな調整もPC版と同じように可能です。",
-  },
-  {
-    q: "未提出者への自動リマインドはできますか？",
-    a: "できます。未提出者に自動でリマインドを送り、催促漏れを減らします。",
-  },
-  {
-    q: "Excelや紙のシフト管理から移行できますか？",
-    a: "できます。希望シフトの回収から確定の連絡までをシフトリに移すと、Excelへの転記や個別連絡が減ります。",
-  },
-];
