+++ src/pages/faq/index.tsx
@@ -1,5 +1,6 @@
 import { Box, Container, Link } from "@chakra-ui/react";
-import { FaqSection, FooterSection } from "@/src/components/features/LandingPage";
+import { FaqSection } from "@/src/components/features/LandingPage/FaqSection";
+import { FooterSection } from "@/src/components/features/NewLandingPage/FooterSection";
 import { HEADER_HEIGHT, Header } from "@/src/components/templates/Header";
 
 export function FaqPage() {
